+++ profile_bluesky/startup/instrument/devices/__init__.py
@@ -7,6 +7,8 @@ from .permit import *
 
 from .constants import *
 from .general_terms import *
+from .sample_data import *
+from .user_data import *
 
 from .amplifiers import *
 from .autosave import *
@@ -18,7 +20,6 @@ from .linkam import *
 from .miscellaneous import *
 from .monochromator import *
 from .protection_plc import *
-from .sample_data import *
 from .scaler import *
 from .shutters import *
 from .slits import *
@@ -26,7 +27,6 @@ from .stages import *
 from .struck3820 import *
 from .suspenders import *
 from .trajectories import *
-from .user_data import *
 
 from .alta import *
 from .blackfly import *
