+++ profile_bluesky/startup/instrument/plans/__init__.py
@@ -2,3 +2,7 @@
 """
 local, custom Bluesky plans (scans) and other functions
 """
+
+from .axis_tuning import *
+from .requested_stop import *
+from .mode_changes import *
