+++ profile_bluesky/startup/instrument/plans/mode_changes.py
@@ -1,0 +1,13 @@
+
+"""
+Support the different instrument modes
+"""
+
+__all__ = [
+    'mode_USAXS',
+    ]
+
+from ..session_logs import logger
+logger.info(__file__)
+
+def mode_USAXS(): ...  # TODO:
