+++ profile_bluesky/startup/instrument/plans/requested_stop.py
@@ -1,0 +1,16 @@
+
+"""
+User can Set a PV to request scanning to stop
+
+Scanning will stop between scans at next loop through scan sequence.
+"""
+
+__all__ = [
+    'IfRequestedStopBeforeNextScan', 
+    # 'undulator',
+    ]
+
+from ..session_logs import logger
+logger.info(__file__)
+
+def IfRequestedStopBeforeNextScan(): ...  # TODO: 41-commands
